KVM: PPC: Book3S HV: Align gfn to L1 page size when inserting nest-rmap entry
BugLink: https://canonical.my.salesforce.com/5003z00001xnYEC (KVM: Back port POWER9 nested virtualization patches (SF: #00228446))

Nested rmap entries are used to store the translation from L1 gpa to L2
gpa when entries are inserted into the shadow (nested) page tables. This
rmap list is located by indexing the rmap array in the memslot by L1
gfn. When we come to search for these entries we only know the L1 page size
(which could be PAGE_SIZE, 2M or a 1G page) and so can only select a gfn
aligned to that size. This means that when we insert the entry, so we can
find it later, we need to align the gfn we use to select the rmap list
in which to insert the entry to L1 page size as well.

By not doing this we were missing nested rmap entries when modifying L1
ptes which were for a page also passed through to an L2 guest.

Signed-off-by: Suraj Jitindar Singh <sjitindarsingh@gmail.com>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
Signed-off-by: Leonardo Garcia <lagarcia@br.ibm.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>