UBUNTU: [Config] ibm-gt: Set CONFIG_DRM_HISI_HIBMC to arm64 only
BugLink: https://bugs.launchpad.net/bugs/1762940

Hisilicon say that the hibmc_drm driver is for use on arm64 arch only,
and is not meant for amd64.

When hibmc_drm is used with amd64 hardware, multiple issues occur which
lead to the screen being unreadable, most significant is the inability
to use the d-i server installer due to the problem.

This patch removes CONFIG_DRM_HISI_HIBMC from all architectures other
than arm64.

[juergh: Picked from Bionic master.]
Signed-off-by: Juerg Haefliger <juergh@canonical.com>